converted prices to decimal

+++ src/Service.PriceHistory.Domain.Models/InstrumentPriceRecord.cs
@@ -6,8 +6,8 @@ namespace Service.PriceHistory.Domain.Models
     {
         public string BrokerId { get; set; }
         public string InstrumentSymbol { get; set; }
-        public double CurrentPrice { get; set; }
-        public double H24P { get; set; }
+        public decimal CurrentPrice { get; set; }
+        public decimal H24P { get; set; }
         public BasePrice H24 { get; set; }
         public BasePrice D7 { get; set; }
         public BasePrice M1 { get; set; }
@@ -16,7 +16,7 @@ namespace Service.PriceHistory.Domain.Models
 
     public class BasePrice
     {
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public DateTime RecordTime { get; set; }
     }
 }
